Implemented an improved lambda conversion.
The calculations used in the example code should only be considered when operating at lambda levels near λ 1. Using this calculation approximation outside the intended area of operation will give a proportional error. The calculation should be used for educational purposes and the accuracy is only adequate when operating near λ 1. When measuring at much leaner or richer gases an improved method of converting the CJ125 output voltage is required. With this updated code, a lookup table is implemented to correct the described conversion error.